Fixed log messages and minor changes
