+++ src/main/java/eu/proteus/consumer/utils/ConsumerUtils.java
@@ -1,4 +1,4 @@
-package eu.proteus.producer.utils;
+package eu.proteus.consumer.utils;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -6,7 +6,8 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
 
-import eu.proteus.consumer.utils.KafkaTopics;
+import com.couchbase.client.java.Bucket;
+import com.couchbase.client.java.Cluster;
 
 public class ConsumerUtils {
 
@@ -35,6 +36,15 @@ public class ConsumerUtils {
 		if (name.equals(KafkaTopics.SIMPLE_MOMENTS.toString()))
 			ret = "simple-moments";
 		return ret;
+	}
+
+	public static Bucket selectCouchbaseBucket(Cluster cluster, String topic) {
+		String myBucket = "proteus";
+		for (ProteusBuckets bucket : ProteusBuckets.values()) {
+			if (!topic.contains("PROTEUS"))
+				myBucket = bucket.name();
+		}
+		return cluster.openBucket(myBucket.toLowerCase());
 	}
 
 }
